Initial work on embed filter
