+++ assets/js/src/shortcake-bakery-admin.js
@@ -111,41 +111,5 @@ jQuery( document ).ready( function ( $ ) {
 			wp.media.editor.open( editor, options );
 		});
 
-	/**
-	 * Attach JS hooks for shortcodes that need them.
-	 *
-	 */
-	if ( typeof wp !== 'undefined' &&
-		 typeof wp.shortcake !== 'undefined' &&
-		 typeof wp.shortcake.hooks !== 'undefined' ) {
-
-		/* Optional fields for GoogleDocs shortcode, displayed conditionally depending on the "type" field */
-		var gdocUrlField = [ ShortcakeBakery.shortcodes['Shortcake_Bakery\\Shortcodes\\GoogleDocs'], 'url' ].join('.');
-		var gdocFields = {
-			all:          [ 'headers', 'start', 'loop', 'delayms' ],
-			spreadsheets: [ 'headers' ],
-			presentation: [ 'start', 'loop', 'delayms' ]
-		};
-
-		wp.shortcake.hooks.addAction( gdocUrlField, function( changed, collection, shortcode ) {
-			if ( 'undefined' === typeof changed.value ) {
-				return;
-			}
-
-			var docUrl = changed.value,
-				docUrlParts = docUrl.split('/'),
-				docType = docUrlParts.length > 3 ? docUrlParts[3] : false;
-
-			_.each( gdocFields.all, function( fieldname ) {
-				var field = sui.views.editAttributeField.getField( collection, fieldname );
-				if ( 'undefined' !== typeof gdocFields[ docType ] && _.contains( gdocFields[ docType ], fieldname ) ) {
-					field.$el.show()
-				} else {
-					field.$el.hide();
-				}
-			} );
-		 } );
-	}
-
 });
